requirements.txt
alembic==1.8.1
asgiref==3.5.2
astroid==2.12.12
asyncmy==0.2.5
attrs==22.1.0
autopep8==1.7.0
bcrypt==4.0.1
Cerberus==1.3.4
cfgv==3.3.1
click==8.1.3
dill==0.3.5.1
distlib==0.3.6
Faker==15.1.1
filelock==3.8.0
Flask==2.2.2
greenlet==1.1.3.post0
identify==2.5.6
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.3
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
py==1.11.0
pycodestyle==2.9.1
PyJWT==2.6.0
pylint==2.15.4
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-asyncio==0.20.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.42
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
virtualenv==20.16.5
Werkzeug==2.2.2
wrapt==1.14.1
